switch to fixed versions
Buildroot doesn't support branch names for the versions. Switch to fixed
versions and add a mechanism to automatically set the version of a
package to the latest HEAD of a given branch. This only works if the
package fetches its source from a git repository.

Set USE_HEAD_VERSION to "y" to build the lastest development versions of
linux and uboot:
  USE_HEAD_VERSION=y make

Co-developed-by: Heiko Thiery <heiko.thiery@kontron.com>
Signed-off-by: Heiko Thiery <heiko.thiery@kontron.com>
Signed-off-by: Michael Walle <michael.walle@kontron.com>