add extlinux.conf to boot partition
An partition UUID is added to the rootfs partition. This partition UUID will
be passed to linux by the extlinux to select the rootfs.

To have UUID partition ID support we switch to GPT partition table type.
Since the TI SoC seems not to support booting from GTP partition we need
ti switch to 'hybrid' mode (see [1] for details).

[1] https://github.com/pengutronix/genimage?tab=readme-ov-file#hdimage

Signed-off-by: Heiko Thiery <heiko.thiery@kontron.com>